pyproject.toml
[tool.poetry]
# Не забудьте заполнить метаданные для вашего проекта!
# Это поможет `poetry` правильно оперировать ими при
# публикации проекта на PyPI и т. д.
name = "simple-bot"
version = "0.1.0"
description = ""
authors = []
license = ""

[tool.poetry.dependencies]
# Здесь находятся зависимости, которые необходимы
# для работы проекта. Без него исходный код не заработает.
python = "^3.9"
vkbottle = "^3.0.2"
envparse = "^0.2.0"
tortoise-orm = "^0.17.8"

[tool.poetry.dev-dependencies]
# Сюда попадают зависимости, помеченные флагом `--dev` -
# то есть те, которые используются при непосредственной разработке.
# Они устанавливаются и работают точно так же, как и обычные зависимости,
# но между ними есть разграничение и при желании установку таких зависимостей
# можно отключить флагом `--no-dev` при установке.

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
